Move phone number last. Next step will be to ask for an opt code.
Please note that we cannot set no password, so most likely I will generate
a random one on the ui. This is not great. Maybe we could further update
the db to make sure that we randomize the password, but for the moment
this should be ok
