+++ hooks/useRegistrationForm.ts
@@ -1,8 +1,8 @@
-import {useState} from "react";
-import {FormData, ValidationErrors} from "@/types/form";
-import {validateAuthFields, validateUserFields} from "@/utils/validation";
-import {dateRange, payTaxToOptions} from "@/lib/constants";
-import {toast} from "sonner";
+import { useState } from "react";
+import { FormData, ValidationErrors } from "@/types/form";
+import { validateUserFields } from "@/utils/validation"; // Removed validateAuthFields
+import { dateRange, payTaxToOptions } from "@/lib/constants";
+import { toast } from "sonner";
 import {supabaseBrowserClient} from "@/lib/supabase/client";
 import { useRouter} from "next/navigation";
 import {getNewUserMetadata} from "@/lib/supabase/database/user";
@@ -32,9 +32,9 @@ const initialFormData: FormData = {
 };
 
 const initialValidationErrors: ValidationErrors = {
-  email: "",
-  password: "",
-  confirmPassword: "",
+  // email: "", // Removed
+  // password: "", // Removed
+  // confirmPassword: "", // Removed
   name: "",
   age: "",
   phone: "",
@@ -82,26 +82,56 @@ export function useRegistrationForm() {
   const validateStep = (step: number): boolean => {
     let errors: ValidationErrors = {};
 
+    // Step 1 (Personal Details) validation
     if (step === 1) {
-      errors = validateAuthFields(formData.authData);
-    } else if (step === 2) {
-      errors = validateUserFields(formData.userData);
+      // Validate name and age from userData
+      const nameError = validateUserFields({ name: formData.userData.name } as FormData["userData"]).name;
+      const ageError = validateUserFields({ age: formData.userData.age } as FormData["userData"]).age;
+
+      if (nameError) errors.name = nameError;
+      if (ageError) errors.age = ageError;
+
+    }
+    // Step 2 (Registration Details) validation
+    else if (step === 2) {
+      // Validate the rest of userData fields excluding name, age, phone, imageUrl
+      const { name, age, phone, imageUrl, ...registrationData } = formData.userData;
+      errors = validateUserFields(registrationData as FormData["userData"]); // Validate remaining fields
+    }
+    // Step 3 (Phone & Confirmation) validation
+    else if (step === 3) {
+        // Validate phone from userData
+        const phoneError = validateUserFields({ phone: formData.userData.phone } as FormData["userData"]).phone;
+        if (phoneError) errors.phone = phoneError;
     }
 
+
     setValidationErrors(errors);
-    return Object.keys(errors).every(
+    // Check if the relevant error keys for the current step are empty
+    const relevantErrorKeys = step === 1 ? ['name', 'age', 'image'] :
+                              step === 2 ? Object.keys(initialValidationErrors).filter(k => !['name', 'age', 'phone', 'image'].includes(k)) :
+                              step === 3 ? ['phone'] : [];
+
+    return relevantErrorKeys.every(
       (key) => !errors[key as keyof ValidationErrors]
     );
   };
 
   const handleNext = () => {
+    // Validate the current step before proceeding
     if (validateStep(step)) {
       setStep((prev) => prev + 1);
-      setValidationErrors(initialValidationErrors);
+      setValidationErrors(initialValidationErrors); // Clear errors for the next step
     } else {
-      toast.error("Te rugăm să completezi toate câmpurile corect", {
-        description: "Verifică câmpurile marcate cu roșu și încearcă din nou.",
-        duration: 4000,
+      // Construct a more specific error message based on the step
+      let errorMessage = "Te rugăm să completezi toate câmpurile obligatorii corect";
+      if (step === 1) errorMessage = "Verifică numele, vârsta și poza.";
+      else if (step === 2) errorMessage = "Verifică detaliile înregistrării.";
+      else if (step === 3) errorMessage = "Verifică numărul de telefon.";
+
+      toast.error(errorMessage, {
+        description: "Verifică câmpurile marcate și încearcă din nou.",
+        duration: 5000,
       });
     }
   };
@@ -112,8 +142,29 @@ export function useRegistrationForm() {
   };
 
     const handleSubmit = async () => {
+        // First, validate Step 3 fields (phone and agreement)
+        if (!validateStep(3) || !agreementChecked) {
+             if (!agreementChecked) {
+                toast.error("Trebuie să fii de acord cu regulamentul", {
+                    description: "Bifează căsuța de acord înainte de a continua.",
+                    duration: 4000,
+                });
+             } else {
+                 toast.error("Te rugăm să completezi numărul de telefon corect", {
+                    description: "Verifică câmpul marcat și încearcă din nou.",
+                    duration: 4000,
+                 });
+             }
+            return; // Stop submission if validation fails or agreement not checked
+        }
+
         setIsLoading(true);
         const currentEdition = await getActiveEdition(); // TODO remove this after we have edition selector on the UI
+        if (!currentEdition) {
+            toast.error("Nu am putut găsi ediția activă a taberei.");
+            setIsLoading(false);
+            return;
+        }
         const metaData = getNewUserMetadata(formData, currentEdition.id);
         // Create auth user with metadata. See `create_user_profiles_table` to understand
         const {error, data: {user}} = await supabaseBrowserClient.auth.signUp({
@@ -127,34 +178,12 @@ export function useRegistrationForm() {
 
         setIsLoading(false);
         if (error) {
-            if (error.code === "user_already_exists") {
-                toast.error("Adresa de email există deja", {
-                    description:
-                        "Dacă ai deja cont, apasă pe butonul de conectare. Dacă nu, folosește altă adresă de email.",
-                    duration: 5000,
-                    action: {
-                        label: "Conectare",
-                        onClick: () => router.push("/login"),
-                    },
-                });
-                return;
-            }
-            if (error.code === "validation_failed") {
-                toast.error("Adresa de email nu este validă.", {
-                    description: "Te rugăm să verifici adresa introdusă.",
-                });
-                return;
-            }
-            if (error.code === "weak_password") {
-                toast.error("Parola este prea slabă.", {
-                    description: "Te rugăm să alegi o parolă mai puternică. " + error.message,
-                });
-                return;
-            }
-
-            toast.error("Ceva nu a mers bine", {
-                description: error.message,
-            });
+            console.error("Sign up error:", error);
+            // Handle specific errors (e.g., phone number already exists if using phone auth later)
+            // For now, a generic error message
+             toast.error("Eroare la crearea contului", {
+                 description: error.message || "Te rugăm să încerci din nou sau contactează suportul.",
+             });
             return;
         }
